+++ src/Products/PloneMeeting/browser/itemcompleteness.py
@@ -2,10 +2,10 @@
 
 from AccessControl import Unauthorized
 from DateTime import DateTime
+from imio.helpers.cache import get_current_user_id
 from plone import api
 from Products.Archetypes import DisplayList
 from Products.Five.browser import BrowserView
-from Products.PloneMeeting.utils import get_current_user_id
 
 
 class ItemCompletenessView(BrowserView):